chore(wfc-rewrite): clean up old wfc code
