--- src/roboarena/server/level_generation/test.json
@@ -1,1 +1,0 @@
-[(Vector(x=0, y=0), Vector(x=-1, y=0), (['┼', '┴', '┬', '├', '─'], ['┼'], ['┼', '┴', '┬', '┤', '─'], ['┼'])), (Vector(x=-1, y=-1), Vector(x=-1, y=0), (['┼', '┴', '┬', '├', '─'], ['┼', '┴', '┬', '┤', '├', '─', '│'], ['┼', '┴', '┬', '┤', '├', '─', '│'], ['┼', '┴', '┬', '┤', '├', '─', '│'])), (Vector(x=-1, y=1), Vector(x=-1, y=0), (['┼', '┴', '┬', '├', '─'], ['┼', '┴', '┬', '┤', '├', '─', '│'], ['┼', '┴', '┬', '┤', '├', '─', '│'], ['┼', '┴', '┬', '┤', '├', '─', '│'])), (Vector(x=-2, y=0), Vector(x=-1, y=0), (['┼', '┴', '┬', '├', '─'], ['┼', '┴', '┬', '┤', '├', '─', '│'], ['┼', '┴', '┬', '┤', '├', '─', '│'], ['┼', '┴', '┬', '┤', '├', '─', '│'])), (Vector(x=0, y=0), Vector(x=-1, y=0), (['┼', '┴', '├'], ['┼'], ['┼', '┴', '┬', '┤', '─'], ['┼'])), (Vector(x=-1, y=-1), Vector(x=-1, y=0), (['┼', '┴', '├'], ['┤', '│'], ['┼', '┬', '┤', '├', '│'], ['┤', '│'])), (Vector(x=-1, y=1), Vector(x=-1, y=0), (['┼', '┴', '├'], ['┼', '┴', '┬', '┤', '├', '─', '│'], ['┼', '┴', '┬', '┤', '├', '─', '│'], ['┼', '┴', '┬', '┤', '├', '─', '│'])), (Vector(x=-2, y=0), Vector(x=-1, y=0), (['┼', '┴', '├'], ['┼', '┴', '┬', '┤', '├', '─', '│'], ['┼', '┴', '┬', '┤', '├', '─', '│'], ['┼', '┴', '┬', '┤', '├', '─', '│']))]
